refactor(usage): remove legacy org usage details UI
