+++ apps/web/src/components/usage-analytics/AIAdoptionDistribution.tsx
