+++ apps/web/src/components/usage-analytics/BackgroundChart.tsx
