+++ apps/web/src/routers/organizations/organization-usage-details-router.ts
@@ -1,5 +1,3 @@
-import type { UsageDetails } from '@/lib/organizations/organization-types';
-import { TimePeriodSchema } from '@/lib/organizations/organization-types';
 import { createTRPCRouter } from '@/lib/trpc/init';
 import {
   OrganizationIdInputSchema,
@@ -7,16 +5,9 @@ import {
   organizationOwnerMutationProcedure,
 } from '@/routers/organizations/utils';
 import { db, readDb } from '@/lib/drizzle';
-import { timedUsageQuery } from '@/lib/usage-query';
-import {
-  microdollar_usage,
-  kilocode_users,
-  organizations,
-  organization_recommendation_dismissals,
-} from '@kilocode/db/schema';
-import { eq, sum, count, sql, and, gte, lte, isNull } from 'drizzle-orm';
+import { organizations, organization_recommendation_dismissals } from '@kilocode/db/schema';
+import { eq, and, isNull } from 'drizzle-orm';
 import * as z from 'zod';
-import { AUTOCOMPLETE_MODEL } from '@/lib/constants';
 import {
   FEATURE_ADOPTION_KEYS,
   getOrganizationFeatureAdoption,
@@ -41,56 +32,9 @@ import {
   calculateWeeklyTrends,
 } from '@/lib/organizations/ai-adoption-calculations';
 
-const UsageDetailsInputSchema = OrganizationIdInputSchema.extend({
-  period: TimePeriodSchema.default('month'),
-  userFilter: z.enum(['all', 'me']).default('all'),
-  groupByModel: z.boolean().default(false),
-});
-
 const UsageTimeseriesInputSchema = OrganizationIdInputSchema.extend({
   startDate: z.iso.datetime(),
   endDate: z.iso.datetime(),
-});
-
-const UsageTimeseriesOutputSchema = z.object({
-  timeseries: z.array(
-    z.object({
-      datetime: z.string().datetime(),
-      name: z.string(),
-      email: z.string(),
-      model: z.string(),
-      provider: z.string(),
-      projectId: z.string().nullable(),
-      costMicrodollars: z.number(),
-      inputTokenCount: z.number(),
-      outputTokenCount: z.number(),
-      requestCount: z.number(),
-    })
-  ),
-});
-
-const UsageDetailsResponseSchema = z.object({
-  daily: z.array(
-    z.object({
-      date: z.string(),
-      user: z.object({
-        name: z.string(),
-        email: z.string(),
-      }),
-      model: z.string().optional(),
-      microdollarCost: z.string().nullable(),
-      tokenCount: z.number(),
-      inputTokens: z.number(),
-      outputTokens: z.number(),
-      requestCount: z.number(),
-    })
-  ),
-});
-
-const AutocompleteMetricsOutputSchema = z.object({
-  cost: z.number(),
-  requests: z.number(),
-  tokens: z.number(),
 });
 
 const FeatureAdoptionOutputSchema = z.object({
@@ -182,26 +126,6 @@ const AIAdoptionTimeseriesOutputSchema = z.object({
   isNewOrganization: z.boolean(), // True if first activity was < 3 days ago
 });
 
-function daysAgo(days: number): string {
-  const now = new Date();
-  return new Date(now.getTime() - days * 24 * 60 * 60 * 1000).toISOString();
-}
-
-function getDateThreshold(period: string): string | null {
-  switch (period) {
-    case 'week':
-      return daysAgo(7);
-    case 'month':
-      return daysAgo(30);
-    case 'year':
-      return daysAgo(365);
-    case 'all':
-      return null; // No date filtering for "all"
-    default:
-      return daysAgo(7); // Default to week
-  }
-}
-
 export const organizationsUsageDetailsRouter = createTRPCRouter({
   getFeatureAdoption: organizationMemberProcedure
     .input(OrganizationIdInputSchema)
@@ -264,278 +188,6 @@ export const organizationsUsageDetailsRouter = createTRPCRouter({
           )
         );
       return { restored: true };
-    }),
-  getTimeSeries: organizationMemberProcedure
-    .input(UsageTimeseriesInputSchema)
-    .output(UsageTimeseriesOutputSchema)
-    .query(async ({ input }) => {
-      const { organizationId, startDate, endDate } = input;
-
-      // Calculate if range is > 90 days to determine grouping strategy
-      const start = new Date(startDate);
-      const end = new Date(endDate);
-      const daysDiff = (end.getTime() - start.getTime()) / (1000 * 60 * 60 * 24);
-      const groupByDay = daysDiff > 90;
-
-      // Create the time bucket SQL based on grouping strategy
-      const timeBucket = groupByDay
-        ? sql<string>`DATE_TRUNC('day', ${microdollar_usage.created_at})`
-        : sql<string>`DATE_TRUNC('hour', ${microdollar_usage.created_at})`;
-
-      // First, get the aggregated usage data
-      const usageData = await timedUsageQuery(
-        {
-          db: readDb,
-          route: 'organizations.usageDetails.getTimeSeries',
-          queryLabel: 'org_usage_timeseries',
-          scope: 'org',
-          period: `${startDate}/${endDate}`,
-        },
-        tx =>
-          tx
-            .select({
-              datetime: timeBucket.as('datetime'),
-              userName: kilocode_users.google_user_name,
-              userEmail: kilocode_users.google_user_email,
-              model: sql<string>`COALESCE(${microdollar_usage.requested_model}, ${microdollar_usage.model})`,
-              provider: microdollar_usage.provider,
-              projectId: microdollar_usage.project_id,
-              costMicrodollars: sum(microdollar_usage.cost),
-              inputTokenCount: sum(microdollar_usage.input_tokens),
-              outputTokenCount: sum(microdollar_usage.output_tokens),
-              requestCount: count(microdollar_usage.id),
-            })
-            .from(microdollar_usage)
-            .innerJoin(kilocode_users, eq(kilocode_users.id, microdollar_usage.kilo_user_id))
-            .where(
-              and(
-                eq(microdollar_usage.organization_id, organizationId),
-                gte(microdollar_usage.created_at, startDate),
-                lte(microdollar_usage.created_at, endDate)
-              )
-            )
-            .groupBy(
-              timeBucket,
-              kilocode_users.google_user_name,
-              kilocode_users.google_user_email,
-              sql`COALESCE(${microdollar_usage.requested_model}, ${microdollar_usage.model})`,
-              microdollar_usage.provider,
-              microdollar_usage.project_id
-            )
-            .orderBy(timeBucket)
-      );
-
-      // Fill in 0's for missing data
-      // TODO surely there's a better way to do this
-      // Get all unique combinations of user/model/provider/projectId from the data
-      const uniqueCombinations = new Map<
-        string,
-        { name: string; email: string; model: string; provider: string; projectId: string | null }
-      >();
-      usageData.forEach(row => {
-        const key = `${row.userEmail}|${row.model}|${row.provider}|${row.projectId || 'null'}`;
-        if (!uniqueCombinations.has(key)) {
-          uniqueCombinations.set(key, {
-            name: row.userName || 'Unknown',
-            email: row.userEmail || 'unknown@example.com',
-            model: row.model || 'Unknown',
-            provider: row.provider || 'Unknown',
-            projectId: row.projectId || null,
-          });
-        }
-      });
-
-      // Generate all time buckets
-      const allTimeBuckets: Date[] = [];
-      const currentBucket = new Date(start);
-      const endBucket = new Date(end);
-
-      if (groupByDay) {
-        currentBucket.setUTCHours(0, 0, 0, 0);
-        endBucket.setUTCHours(0, 0, 0, 0);
-      } else {
-        currentBucket.setUTCMinutes(0, 0, 0);
-        endBucket.setUTCMinutes(0, 0, 0);
-      }
-
-      while (currentBucket <= endBucket) {
-        allTimeBuckets.push(new Date(currentBucket));
-        if (groupByDay) {
-          currentBucket.setUTCDate(currentBucket.getUTCDate() + 1);
-        } else {
-          currentBucket.setUTCHours(currentBucket.getUTCHours() + 1);
-        }
-      }
-
-      // Create a map of existing data for quick lookup
-      const dataMap = new Map<string, (typeof usageData)[0]>();
-      usageData.forEach(row => {
-        const key = `${new Date(row.datetime).toISOString()}|${row.userEmail}|${row.model}|${row.provider}|${row.projectId || 'null'}`;
-        dataMap.set(key, row);
-      });
-
-      // Generate complete timeseries with zeros for missing data
-      const timeseries: Array<{
-        datetime: string;
-        name: string;
-        email: string;
-        model: string;
-        provider: string;
-        projectId: string | null;
-        costMicrodollars: number;
-        inputTokenCount: number;
-        outputTokenCount: number;
-        requestCount: number;
-      }> = [];
-
-      allTimeBuckets.forEach(bucket => {
-        const bucketISO = bucket.toISOString();
-        uniqueCombinations.forEach(combo => {
-          const key = `${bucketISO}|${combo.email}|${combo.model}|${combo.provider}|${combo.projectId || 'null'}`;
-          const existingData = dataMap.get(key);
-
-          timeseries.push({
-            datetime: bucketISO,
-            name: combo.name,
-            email: combo.email,
-            model: combo.model,
-            provider: combo.provider,
-            projectId: combo.projectId,
-            costMicrodollars: existingData
-              ? Number.parseInt(existingData.costMicrodollars?.toString() || '0')
-              : 0,
-            inputTokenCount: existingData ? Number(existingData.inputTokenCount) || 0 : 0,
-            outputTokenCount: existingData ? Number(existingData.outputTokenCount) || 0 : 0,
-            requestCount: existingData ? Number(existingData.requestCount) || 0 : 0,
-          });
-        });
-      });
-
-      return { timeseries };
-    }),
-  get: organizationMemberProcedure
-    .input(UsageDetailsInputSchema)
-    .output(UsageDetailsResponseSchema)
-    .query(async ({ input, ctx }): Promise<UsageDetails> => {
-      const { organizationId, period, userFilter, groupByModel } = input;
-
-      const dateThreshold = getDateThreshold(period);
-
-      const whereConditions = [eq(microdollar_usage.organization_id, organizationId)];
-      if (dateThreshold) {
-        whereConditions.push(gte(microdollar_usage.created_at, dateThreshold));
-      }
-
-      // Add user filter if "me" is selected
-      if (userFilter === 'me') {
-        whereConditions.push(eq(microdollar_usage.kilo_user_id, ctx.user.id));
-      }
-
-      const usageDetails = await timedUsageQuery(
-        {
-          db: readDb,
-          route: 'organizations.usageDetails.get',
-          queryLabel: 'org_usage_daily',
-          scope: 'org',
-          period,
-        },
-        tx =>
-          tx
-            .select({
-              date: sql<string>`DATE(${microdollar_usage.created_at})`.as('date'),
-              userName: kilocode_users.google_user_name,
-              userEmail: kilocode_users.google_user_email,
-              ...(groupByModel && {
-                model: sql<
-                  string | null
-                >`COALESCE(${microdollar_usage.requested_model}, ${microdollar_usage.model})`,
-              }),
-              microdollarCost: sum(microdollar_usage.cost),
-              tokenCount: sum(
-                sql`${microdollar_usage.input_tokens} + ${microdollar_usage.output_tokens} + ${microdollar_usage.cache_write_tokens} + ${microdollar_usage.cache_hit_tokens}`
-              ),
-              inputTokens: sum(microdollar_usage.input_tokens),
-              outputTokens: sum(microdollar_usage.output_tokens),
-              requestCount: count(microdollar_usage.id),
-            })
-            .from(microdollar_usage)
-            .innerJoin(kilocode_users, eq(kilocode_users.id, microdollar_usage.kilo_user_id))
-            .where(and(...whereConditions))
-            .groupBy(
-              sql`DATE(${microdollar_usage.created_at})`,
-              kilocode_users.google_user_name,
-              kilocode_users.google_user_email,
-              ...(groupByModel
-                ? [sql`COALESCE(${microdollar_usage.requested_model}, ${microdollar_usage.model})`]
-                : [])
-            )
-            .orderBy(sql`DATE(${microdollar_usage.created_at}) DESC`)
-      );
-
-      const daily = usageDetails.map(row => ({
-        date: row.date,
-        user: {
-          name: row.userName,
-          email: row.userEmail,
-        },
-        ...(groupByModel && { model: 'model' in row ? row.model || undefined : undefined }),
-        microdollarCost: row.microdollarCost?.toString() || null,
-        tokenCount: Number(row.tokenCount) || 0,
-        inputTokens: Number(row.inputTokens) || 0,
-        outputTokens: Number(row.outputTokens) || 0,
-        requestCount: Number(row.requestCount) || 0,
-      }));
-
-      return {
-        daily,
-      };
-    }),
-  getAutocomplete: organizationMemberProcedure
-    .input(
-      OrganizationIdInputSchema.extend({
-        period: TimePeriodSchema.default('month'),
-      })
-    )
-    .output(AutocompleteMetricsOutputSchema)
-    .query(async ({ input }) => {
-      const { organizationId, period } = input;
-
-      const dateThreshold = getDateThreshold(period);
-
-      const whereConditions = [
-        eq(microdollar_usage.organization_id, organizationId),
-        eq(microdollar_usage.model, AUTOCOMPLETE_MODEL),
-      ];
-      if (dateThreshold) {
-        whereConditions.push(gte(microdollar_usage.created_at, dateThreshold));
-      }
-
-      const result = await timedUsageQuery(
-        {
-          db: readDb,
-          route: 'organizations.usageDetails.getAutocomplete',
-          queryLabel: 'org_autocomplete_aggregate',
-          scope: 'org',
-          period,
-        },
-        tx =>
-          tx
-            .select({
-              total_cost: sql<number>`COALESCE(SUM(${microdollar_usage.cost}), 0)::float`,
-              request_count: sql<number>`COUNT(*)::float`,
-              total_tokens: sql<number>`COALESCE(SUM(${microdollar_usage.input_tokens}) + SUM(${microdollar_usage.output_tokens}), 0)::float`,
-            })
-            .from(microdollar_usage)
-            .where(and(...whereConditions))
-      );
-
-      const metrics = result[0] || { total_cost: 0, request_count: 0, total_tokens: 0 };
-
-      return {
-        cost: metrics.total_cost,
-        requests: metrics.request_count,
-        tokens: metrics.total_tokens,
-      };
     }),
   getAIAdoptionTimeseries: organizationMemberProcedure
     .input(UsageTimeseriesInputSchema)
